back end/requirements.txt
# Python dependencies for BeeAiRating project
# Core dependencies
requests>=2.31.0
flask>=2.3.0
numpy>=1.24.0
pandas>=2.0.0

# Web development
jinja2>=3.1.0
werkzeug>=2.3.0

# Data processing
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0

# AI/ML Services
google-generativeai>=0.3.0

# Web scraping and analysis
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 